backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib
openai
python-multipart
